Repost feature in FeedPage and fixed bug in likes hover
