+++ src/Pages/HashtagPage.js
@@ -2,7 +2,7 @@ import { useEffect, useState, useContext} from "react";
 import { useParams } from "react-router-dom";
 import axios from "axios";
 import { ThreeDots } from "react-loader-spinner";
-import Post from "../Components/Posts/Post";
+import Post from "../Components/Posts/Post/Post";
 import { FeedPage } from "../shared/Feed/FeedPage";
 import UrlContext from "../contexts/UrlContext";
